created abstract class to remove duplicated code
